requirements.txt
scapy==2.5.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
plotext==5.2.8
geoip2==4.7.0
ipwhois==1.2.0
python-whois==0.8.0
joblib==1.3.2
tqdm==4.65.0
pyyaml==6.0.1
colorama==0.4.6
tabulate==0.9.0
